+++ __tests__/lib/user-service.test.ts
@@ -1,24 +1,28 @@
-import { UserService } from "lib/services/implementations/user-service";
-import { SignInErrorType } from "lib/services/interfaces/user-service.interface";
-import { expect, test, vi } from "vitest";
+import { ApiService } from "lib/services/api-service";
+import { SignInErrorType, UserService } from "lib/services/user-service";
+import { beforeEach, expect, test, vi } from "vitest";
+import { anyString, mock, mockReset } from 'vitest-mock-extended';
+
+const mockApiService = mock<ApiService>();
+
+beforeEach(() => {
+    mockReset(mockApiService)
+})
 
 test.concurrent("UserService exists", () =>{
     expect(UserService).toBeDefined();
 });
 
 test.concurrent("UserService has signIn method", () => {
-    const mockApiService = {
-        signIn: vi.fn().mockResolvedValue(true)
-    }
+    mockApiService.signIn.mockReturnValue(new Promise(() => true));
+
     let userService = new UserService(mockApiService);
 
     expect(userService.signIn).toBeInstanceOf(Function);
 });
 
 test.concurrent("UserService.signIn return SignInResult with IsSuccess true when user is successfuly signed in", async () => {
-    const mockApiService = {
-        signIn: vi.fn().mockResolvedValue(true)
-    }
+    mockApiService.signIn.mockReturnValue(new Promise((resolve, reject) => {resolve(true)}));
     let userService = new UserService(mockApiService);
 
     let signInResult = await userService.signIn("testEmail@example.com", "password");
@@ -27,9 +31,7 @@ test.concurrent("UserService.signIn return SignInResult with IsSuccess true when
 })
 
 test.concurrent("UserService.signIn return SignInResult with IsSuccess false when user not exists", async () => {
-    const mockApiService = {
-        signIn: vi.fn().mockResolvedValue(false)
-    }
+    mockApiService.signIn.mockReturnValue(new Promise((resolve, reject) => {resolve(false)}));
     let userService = new UserService(mockApiService);
 
     let signInResult = await userService.signIn("notExistingEmail@example.com", "password");
@@ -38,9 +40,7 @@ test.concurrent("UserService.signIn return SignInResult with IsSuccess false whe
 })
 
 test.concurrent("UserService.sign return SignInResult with ErrorType IncorrectEmailOrPassword when signin failed", async () =>{
-    const mockApiService = {
-        signIn: vi.fn().mockResolvedValue(false)
-    }
+    mockApiService.signIn.mockReturnValue(new Promise((resolve, reject) => {resolve(false)}));
     let userService = new UserService(mockApiService);
 
     let signInResult = await userService.signIn("notExistingEmail@example.com", "password");
@@ -49,9 +49,7 @@ test.concurrent("UserService.sign return SignInResult with ErrorType IncorrectEm
 })
 
 test.concurrent("UserService.sign return SignInResult with ErrorType IncorrectEmailFormat when email format is incorrect", async () =>{
-    const mockApiService = {
-        signIn: vi.fn().mockResolvedValue(false)
-    }
+    mockApiService.signIn.mockReturnValue(new Promise((resolve, reject) => {resolve(false)}));
     let userService = new UserService(mockApiService);
 
     let signInResult = await userService.signIn("incorrectEmail", "password");
