--- src/lib/services/interfaces/api-service.interface.ts
@@ -1,5 +1,0 @@
-export const IApiServiceToken = Symbol("IApiService");
-
-export interface IApiService{
-    signIn(email: string, password: string) : Promise<boolean>;
-}
