--- src/lib/services/interfaces/config-service.interface.ts
@@ -1,5 +1,0 @@
-export const IConfigServiceToken = Symbol("IConfigService");
-
-export interface IConfigService {
-  getApiUrl(): string;
-}
